builds a node to be base level ready
This will install, configure, and run puppet agent to connect to master. Docker and docker-compose are also installed so that any builds that require them will be able to instantly build. Great to be used with other scripts that pull a git repo that has everything ready to docker-compose up -d.